requirements.txt
# PyTorch & vision
torch>=2.5.0
torchvision>=0.20.0

# Image I/O & processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data handling
numpy>=1.23.0
pandas>=1.5.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# Optionally for metrics / train/val split / model persistence
scikit-learn>=1.2.0
tensorboard>=2.12.0

# Jupyter support
notebook>=6.5.0

matplotlib
pytorch-msssim
